Remove old Giga Gravitation description from skill.hsp
When Giga Gravitation's description was changed, the new description was appended instead of overwriting the old one. In other words, the description is first set to the old one, and then to the new one. Probably an old copy/paste error from way back when that was never noticed because it doesn't cause any problems other than wasting a couple processor cycles.